alright, time to do the configs now!
