+++ apps/admin/package.json
@@ -1,0 +1,27 @@
+{
+    "name": "@kv-attendance-app/admin",
+    "version": "1.0.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev --turbopack",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint"
+    },
+    "dependencies": {
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0",
+        "next": "15.1.6"
+    },
+    "devDependencies": {
+        "typescript": "^5",
+        "@types/node": "^20",
+        "@types/react": "^19",
+        "@types/react-dom": "^19",
+        "postcss": "^8",
+        "tailwindcss": "^3.4.1",
+        "eslint": "^9",
+        "eslint-config-next": "15.1.6",
+        "@eslint/eslintrc": "^3"
+    }
+}
