+++ apps/mobile/.easignore
