+++ apps/mobile/.eslintignore
