+++ apps/mobile/app.config.js
