+++ apps/mobile/assets/images/adaptive-icon.png
