+++ apps/mobile/assets/images/icon.png
