+++ apps/mobile/assets/images/splash-icon.png
