+++ apps/mobile/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@kv-attendance-app/frontend",
+    "name": "@kv-attendance-app/mobile",
     "main": "expo-router/entry",
     "version": "1.0.0",
     "scripts": {
