+++ apps/mobile/src/app/(auth)/_layout.tsx
