+++ apps/mobile/src/app/(auth)/sign-up.tsx
