+++ apps/mobile/src/app/(home)/_layout.tsx
