+++ apps/mobile/src/app/(home)/confirm.tsx
