+++ package.json
@@ -1,5 +1,6 @@
 {
     "name": "@kv-attendance-app/repo",
+    "version": "1.0.0",
     "private": true,
     "scripts": {
         "dev": "turbo dev",
